requirements.txt
absl-py==1.4.0
accelerate==0.5.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
backcall==0.2.0
bertviz==1.2.0
blis==0.7.9
boto3==1.26.162
botocore==1.29.162
Brotli==1.0.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmaes==0.9.1
cmake==3.26.4
coloredlogs==15.0.1
colorlog==6.7.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
datasets==1.16.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
editdistance==0.6.0
executing==1.2.0
fairscale==0.4.13
fastDamerauLevenshtein==1.0.7
filelock==3.12.2
fire==0.5.0
flatbuffers==2.0.7
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.0
h5py==3.9.0
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflate64==0.3.1
ipykernel==6.23.3
ipython
ipywidgets==8.0.6
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonlines==3.1.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
lit==16.0.6
llvmlite==0.40.1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
multivolumefile==0.2.3
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.1
neuspell==1.0.0
nlpaug==1.1.7
nltk==3.6.6
numba==0.57.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.14.0
onnxconverter-common==1.13.0
onnxruntime==1.15.1
onnxruntime-tools==1.7.0
opt-einsum==3.3.0
optuna==3.2.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
pooch==1.6.0
portalocker==2.0.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3nvml==0.2.7
py7zr==0.20.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycparser==2.21
pycryptodomex==3.18.0
Pygments==2.15.1
pynndescent==0.5.10
pyparsing==3.1.0
pyppmd==1.0.0
pyspellchecker==0.7.2
python-dateutil==2.8.2
pytorch-pretrained-bert==0.6.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
pyzstd==0.15.9
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rouge-score==0.0.4
rsa==4.9
s3transfer==0.6.1
sacrebleu==1.5.1
sacremoses==0.0.53
scikit-learn==1.2.2
scikit-multilearn==0.2.0
scipy==1.11.0
seaborn==0.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
smart-open==6.3.0
soundfile==0.12.1
soxr==0.3.5
srsly==2.4.6
stack-data==0.6.2
sympy==1.12
termcolor==2.3.0
texttable==1.6.7
tf2onnx==1.14.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.16.2
triton==2.0.0
typer==0.9.0
typing_extensions
tzdata==2023.3
umap-learn==0.5.1
urllib3==1.26.16
wasabi==1.1.2
wcwidth==0.2.6
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wrapt==1.14.1
xmltodict==0.13.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
aiosignal==1.3.1
anthropic==0.20.0
backoff==2.2.1
fastavro==1.8.2
google-cloud-aiplatform==1.38.1
mistralai==0.1.8
openai==1.6.1
